+++ WormGame/Entities/Blocks.cs
@@ -4,11 +4,12 @@ using Otter.Graphics.Drawables;
 using WormGame.Core;
 using WormGame.Static;
 using WormGame.Pooling;
+using WormGame.Entities;
 
 namespace WormGame.Entities
 {
     /// @author Antti Harju
-    /// @version 15.08.2020
+    /// @version 14.08.2020
     /// <summary>
     /// A custom pooler that can be called to create blocks. Uses tilemap to render objects.
     /// </summary>
@@ -62,7 +63,7 @@ namespace WormGame.Entities
 
 
         /// <summary>
-        /// Clears tilemap and resets block module pooler.
+        /// Reset pooler and clear tilemap.
         /// </summary>
         public override void Reset()
         {
@@ -133,8 +134,6 @@ namespace WormGame.Entities
         /// <returns>Block</returns>
         public BlockModule Spawn(Worm worm, Blocks manager)
         {
-            return null;
-            /*
             color = worm.Color;
             bottom = height;
             left = width;
@@ -172,7 +171,7 @@ namespace WormGame.Entities
             if (visualize)
                 Visualize();
 #endif
-            return firstModule;*/
+            return firstModule;
         }
 
 
@@ -277,7 +276,7 @@ namespace WormGame.Entities
         /// <param name="wormLength">Worms length</param>
         /// <returns>Abort spawn process</returns>
         private bool InitBuffer(WormModule wormModule, int wormLength)
-        {/*
+        {
             for (int i = 0; i < wormLength; i++)
             {
                 int x = collision.X(wormModule.Target.X);
@@ -289,7 +288,7 @@ namespace WormGame.Entities
                 if (x > right) right = x;
                 if (y > top) top = y;
                 wormModule = wormModule.Next;
-            }*/
+            }
             return false;
         }
 
@@ -301,12 +300,11 @@ namespace WormGame.Entities
         /// <param name="wormLength">Worms length</param>
         private void ClearBuffer(WormModule wormModule, int wormLength)
         {
-            /*
             for (int i = 0; i < wormLength; i++)
             {
                 optimizationBuffer[collision.X(wormModule.Target.X), collision.Y(wormModule.Target.Y)] = false;
                 wormModule = wormModule.Next;
-            }*/
+            }
         }